Refactor Discovery plugin
